trying to fix view of categories

+++ app/Http/Controllers/Admin/PostController.php
@@ -3,9 +3,11 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Model\Post;
+
 use App\Http\Controllers\Controller;
-use Illuminate\Support\Str;
+use App\Model\Category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -22,13 +24,27 @@ class PostController extends Controller
     }
 
     /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function indexUser()
+    {
+        $posts = Post::where('user_id', Auth::user()->id)->orderBy('created_at', 'desc')->paginate(20);
+
+        return view('admin.posts.index', ['posts' => $posts]);
+    }
+
+    /**
      * Show the form for creating a new resource.
      *
      * @return \Illuminate\Http\Response
      */
     public function create()
     {
-        return view('admin.posts.create');
+        //passiamo le categorie alla pagina create
+        $categories = Category::all();
+        return view('admin.posts.create', ['categories' => $categories]);
     }
 
     /**
@@ -41,12 +57,20 @@ class PostController extends Controller
     {
         $data = $request->all();
         $data['user_id'] = Auth::user()->id;
+        // dd(Category::findOrFail($data['category_id']));
+        // dd(Category::where('id', $data['category_id'])->first());
+        // if (Auth::user()->id != $data['user_id']) {
+        //     abort('404');
+        // }
         $postValidate = $request->validate(
             [
                 'title' => 'required|max:240',
                 'content' => 'required',
+                'category_id' => 'exists:App\Model\Category,id'
             ]
         );
+
+
         $post = new Post();
         $post->fill($data);
         $post->slug = $post->createSlug($data['title']);
@@ -54,13 +78,14 @@ class PostController extends Controller
 
         return redirect()->route('admin.posts.show', $post->slug);
     }
+
     /**
      * Display the specified resource.
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-        public function show(Post $post)
+    public function show(Post $post)
     {
         return view('admin.posts.show', ['post' => $post]);
     }
@@ -73,7 +98,12 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
-        dd($post);
+        if (Auth::user()->id != $post->user_id) {
+            abort('403');
+        }
+        $categories = Category::all();
+
+        return view('admin.posts.edit', ['post' => $post, 'categories' => $categories]);
     }
 
     /**
@@ -83,9 +113,36 @@ class PostController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, Post $post)
     {
-        //
+        $data = $request->all();
+        if (Auth::user()->id != $post->user_id) {
+            abort('403');
+        }
+
+
+        $postValidate = $request->validate(
+            [
+                'title' => 'required|max:240',
+                'content' => 'required',
+                'category_id' => 'exists:App\Model\Category,id'
+            ]
+        );
+
+        if ($data['title'] != $post->title) {
+            $post->title = $data['title'];
+            $post->slug = $post->createSlug($data['title']);
+        }
+        if ($data['content'] != $post->content) {
+            $post->content = $data['content'];
+        }
+        if ($data['category_id'] != $post->category_id) {
+            $post->category_id = $data['category_id'];
+        }
+
+        $post->update();
+
+        return redirect()->route('admin.posts.show', $post);
     }
 
     /**
@@ -96,6 +153,10 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if (Auth::user()->id != $post->user_id) {
+            abort('403');
+        }
+
         $post->delete();
 
         return redirect()->route('admin.posts.index')->with('status', "Post id $post->id deleted");
